Refactor: move config.scada module to actors2.config; move config package to proactor.config
