+++ gw_spaceheat/actors/boolean_actuator.py
@@ -11,7 +11,7 @@ from actors.actor_base import ActorBase
 from actors.utils import QOS
 from actors.utils import Subscription
 from actors.utils import responsive_sleep
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.node_config import NodeConfig
 from data_classes.sh_node import ShNode
