+++ gw_spaceheat/actors/home_alone.py
@@ -10,7 +10,7 @@ from actors.actor_base import ActorBase
 from actors.utils import QOS
 from actors.utils import Subscription
 from actors.utils import responsive_sleep
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.sh_node import ShNode
