+++ gw_spaceheat/actors/scada.py
@@ -36,7 +36,7 @@ from actors.scada_base import ScadaBase
 from actors.utils import QOS
 from actors.utils import Subscription
 from actors.utils import responsive_sleep
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.components.boolean_actuator_component import BooleanActuatorComponent
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.node_config import NodeConfig
