+++ gw_spaceheat/actors/simple_sensor.py
@@ -6,7 +6,7 @@ from gwproto.messages import GtTelemetry_Maker
 from actors.actor_base import ActorBase
 from actors.utils import Subscription
 from actors.utils import responsive_sleep
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.node_config import NodeConfig
 from data_classes.sh_node import ShNode
