+++ gw_spaceheat/actors2/scada_data.py
@@ -21,7 +21,7 @@ from gwproto.messages import SnapshotSpaceheat_Maker
 from gwproto.messages import TelemetrySnapshotSpaceheat
 from gwproto.messages import TelemetrySnapshotSpaceheat_Maker
 
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.node_config import NodeConfig
 from data_classes.sh_node import ShNode
