+++ gw_spaceheat/try_actors.py
@@ -6,7 +6,7 @@ import dotenv
 
 import load_house
 from command_line_utils import run_nodes_main, parse_args
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.sh_node import ShNode
 from schema.enums import Role
