+++ tests/data_classes/test_electric_meter_cac.py
@@ -1,5 +1,5 @@
 import load_house
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.cacs.electric_meter_cac import ElectricMeterCac
 from schema.gt.cacs import GtElectricMeterCac_Maker
