+++ tests/data_classes/test_electric_meter_component.py
@@ -1,5 +1,5 @@
 import load_house
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.components.electric_meter_component import ElectricMeterComponent
 from schema.gt.components import GtElectricMeterComponent_Maker
