+++ tests/test_actors/test_scada.py
@@ -10,7 +10,7 @@ import pytest
 from actors.actor_base import ActorBase
 from actors.scada import Scada
 from actors.scada import ScadaCmdDiagnostic
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.sh_node import ShNode
 from named_tuples.telemetry_tuple import TelemetryTuple
 from gwproto.enums import TelemetryName
