+++ tests/test_misc/test_clean_env.py
@@ -1,7 +1,7 @@
 from tests.conftest import DUMMY_TEST_HARDWARE_LAYOUT_PATH
 
 import pytest
-from config import Paths
+from proactor.config import Paths
 from data_classes.hardware_layout import HardwareLayout
