+++ tests/test_misc/test_flush_and_load_house.py
@@ -3,7 +3,7 @@
 from tests.utils import flush_all
 
 import load_house
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.component import Component
 from data_classes.component_attribute_class import ComponentAttributeClass
 from data_classes.components.boolean_actuator_component import BooleanActuatorCac
