+++ tests/utils/atn_recorder.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from actors.atn import Atn
 from actors.utils import gw_mqtt_topic_decode
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.sh_node import ShNode
 from gwproto.messages import GtShStatus
