+++ tests/utils/scada2_recorder.py
@@ -14,7 +14,7 @@ from gwproto.messages import PingMessage
 from paho.mqtt.client import MQTT_ERR_SUCCESS
 
 from actors2 import Scada2
-from config import ScadaSettings
+from actors2.config import ScadaSettings
 from data_classes.hardware_layout import HardwareLayout
 from data_classes.sh_node import ShNode
 from proactor.message import MQTTReceiptPayload
